feat(TG-290): Added tests for base-chart component
